Convert integration tests into examples
Examples are still executed during `cargo test`,
because test = true has been set in Cargo.toml for each example binary.
