+++ examples/firebase_user.rs
@@ -1,10 +1,10 @@
+use firestore_db_and_auth::Credentials;
 use firestore_db_and_auth::*;
 
 const TEST_USER_ID: &str = include_str!("test_user_id.txt");
 
-#[test]
-fn user_info() -> errors::Result<()> {
-    let cred = credentials::Credentials::from_file("firebase-service-account.json").expect("Read credentials file");
+fn main() -> errors::Result<()> {
+    let cred = Credentials::from_file("firebase-service-account.json").expect("Read credentials file");
 
     let user_session = UserSession::by_user_id(&cred, TEST_USER_ID, false)?;
 
@@ -13,4 +13,9 @@ fn user_info() -> errors::Result<()> {
     assert_eq!(user_info_container.users[0].localId.as_ref().unwrap(), TEST_USER_ID);
 
     Ok(())
+}
+
+#[test]
+fn firebase_user_test() {
+    main().unwrap();
 }
