--- examples/own_auth/readme.md
@@ -1,6 +1,0 @@
-# Example - Own Auth
-
-This example expects a "firebase-service-account.json" file in the working directory.
-
-It shows how to use your own implementation of FirebaseAuthBearer and avoid using any of the `session`
-types.
