--- examples/rocket_http_protected_route/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "rocket_http_protected_route"
-publish = false
-version = "0.0.0"
-authors = ["David Gräff <david.graeff@web.de>"]
-edition = "2018"
-license = "MIT"
-readme = "readme.md"
-
-[dependencies]
-firestore-db-and-auth = { path = "../../", version = "^0", features=["rustls-tls","rocket_support"], default-features=false }
-
-[dependencies.rocket]
-version = "^0.4"
-default-features = false
