--- examples/rocket_http_protected_route/rust-toolchain
@@ -1,1 +1,0 @@
-nightly
